Add user ability condition and project validation
Previously, students could view and apply to all projects. Now, students
can view only projects that are marked as "approved" and can only view
the application link for those they have not applied to. The ability
condition (only students can create submissions) was implemented with
CanCanCan, and the "already applied for this project?" validation was
done in the projects/index.html.erb view (it needs to be moved to
the controller or model).